sync: ask for confirmation when using PAGER
`aur sync` will use $PAGER if `vifm` is not avaiable, falling back to
`less -K` if $PAGER is unset. `aur sync` uses the exit code from this
invocation to decide if the build process should continue or should be
aborted. This is a important part of the process as continuing implies
the user reviewed and agreed to run the PKGBUILD.

While vifm and the current default (less -K) allows exiting with error
by using :cq or ^C respectively, if the user has $PAGER set to something
that doesn't wait for user input and provide a way to exit with error,
or has LESS=F which causes `less` to exit successfully if the output
fits in one page, aur sync will proceed without explicitly consent from
the user.

Let's override the -F option on our default fallback making sure it
always paginate, and prompt for confirmation whenever a custom $PAGER is
set in order to always have explicit consent from the user and a way to
abort if desired.

Fixes #530