completions: bypass root warning when installing
aurutils scripts have a molly-guard that prevents running them as root.

However, the completion generation system needs to call each script with
`--dump-options` to get the list of options to be completed.

To allow installing aur and its completions as root, we disable the
molly-guard setting `AUR_ASROOT=1` on each call.